Move limbs to ZLimb
